Implement double-checked reading from the LruFs.
To promote consistency for ReadWrite blobs manipulated concurrently,
we retry all read operations if the version of the data is not the same
as when the read operation began.

The version therefore behaves like a sequence number in a sequence lock,
and is updated by writes to invalidate pending operations at a previous
version.

This has the nice side effect that the LruFs can ignore versions when
actually reading data, simplifying the impl of Cache for LruFs.