Refactor sparse file handling to use chunk-level locks.
The new strategy effectively stores an FSM in each chunk, where
threads can coordinate using the included Mutex and Condvar to wait
for and cause state transitions to occur.

Currently, the locks used in each chunk are mutual exclusion locks,
but in the future they can be replaced with reader-writer locks, allowing
reads to not block each other.

The included freeze code is preliminary, and must be refactored.