Begin refactoring for new, content-addressing based design..
Unfortunately this commit doesn't really represent a "working"
state and there is some commented out code etc., but we are
at least back to compiling and remaining tests passing.

Most of the new working code is in IndexedSparseFile, then just
refactoring the rest of the code to cooperate with the lifetime
it adds.

Via use of scoped-pool, we actually now also get rid of Arc-ing
the Fs, instead we just stack allocate it and hand out &Fs references
everywhere.